Kaynakları çekmeden önce doğrulama methotu yazıldı. Ek olarak gezegene ait kaynaklar ekrana basılıyor.

+++ WarOfGalaxiesServer/WarOfGalaxiesServer/WarOfGalaxiesApi.DAL/Models/TblUserPlanets.cs
@@ -10,9 +10,9 @@ namespace WarOfGalaxiesApi.DAL.Models
         public string PlanetCordinate { get; set; }
         public int PlanetType { get; set; }
         public string PlanetName { get; set; }
-        public int Metal { get; set; }
-        public int Crystal { get; set; }
-        public int Boron { get; set; }
+        public long Metal { get; set; }
+        public long Crystal { get; set; }
+        public long Boron { get; set; }
         public DateTime LastUpdateDate { get; set; }
     }
 }
